+++ Code/DataStructs/Wrap/testRealValueVect.py
@@ -4,7 +4,7 @@
 
 from __future__ import print_function, unicode_literals
 from rdkit import RDConfig
-import os, sys, pickle
+import os, pickle
 import unittest
 from rdkit import DataStructs as ds