+++ Code/GraphMol/Subgraphs/Subgraphs.cpp
@@ -175,17 +175,6 @@ void recurseWalkRange(
   }
 }
 
-void dumpVIV(VECT_INT_VECT v) {
-  VECT_INT_VECT::iterator i;
-  INT_VECT::iterator j;
-  for (i = v.begin(); i != v.end(); i++) {
-    for (j = i->begin(); j != i->end(); j++) {
-      std::cout << *j << " ";
-    }
-    std::cout << std::endl;
-  }
-}
-
 PATH_LIST
 extendPaths(int *adjMat, unsigned int dim, const PATH_LIST &paths,
             int allowRingClosures = -1, double *distMat = nullptr) {
@@ -463,10 +452,9 @@ findAllPathsOfLengthsMtoN(const ROMol &mol, unsigned int lowerLen,
 
   if (!distMat) {
     // generate the adjacency matrix by hand by looping over the bonds
-    ROMol::ConstBondIterator bondIt;
-    for (bondIt = mol.beginBonds(); bondIt != mol.endBonds(); bondIt++) {
-      Atom *beg = (*bondIt)->getBeginAtom();
-      Atom *end = (*bondIt)->getEndAtom();
+    for (const auto bond : mol.bonds()) {
+      Atom *beg = bond->getBeginAtom();
+      Atom *end = bond->getEndAtom();
       // check for H, which we might be skipping
       if (useHs || (beg->getAtomicNum() != 1 && end->getAtomicNum() != 1)) {
         adjMat[beg->getIdx() * dim + end->getIdx()] = 1;